TRE-680: Add MDC preservation to database operations in repositories for improved logging
