--- app/uk/gov/hmrc/tradereportingextracts/utils/ReportRequestUtil.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright 2025 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.tradereportingextracts.utils
-
-import uk.gov.hmrc.tradereportingextracts.models.ReportRequest
-
-import scala.util.{Failure, Success, Try}
-
-object ReportRequestUtil {
-
-  def isReportStatusComplete(reportRequest: ReportRequest): Boolean =
-    reportRequest.fileNotifications.exists { notifications =>
-      val notificationsCount = notifications.size
-      val lastNotification   = notifications.find(_.reportLastFile == "true")
-      lastNotification match {
-        case Some(last) =>
-          Try(last.reportFilesParts.toInt) match {
-            case Success(parts) => notificationsCount == parts
-            case Failure(_)     => false
-          }
-        case _          => false
-      }
-    }
-}
